% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_imbalance.R
\name{dist_imbalance}
\alias{dist_imbalance}
\title{Imbalance distance caculation}
\usage{
dist_imbalance(obj, base)
}
\arguments{
\item{obj}{The object flow fragment to be forecasted.}

\item{base}{The flow base fragment used by KNN method.}
}
\value{
The distance between object flow and every flow in flow base as a dataframe with
the same order of flow in flow base.
}
\description{
This functions help function knn_flow use imbalance distance in caculating similarity
among traffic flows.
}
\details{
Imbalance distance means that values of flow in flow base is treated differently, which
depends on whether it's higher than object flow or otherwise. Because the punishment of
forecast should be different whether the forecast tends to be higher or smaller. It's more
tolerable that make higher forecast than lower.
}
