% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_euclidean.R
\name{enhanced_euclidean}
\alias{enhanced_euclidean}
\title{Caculate enhanced Euclidean distance}
\usage{
enhanced_euclidean(series_matrix, r1, r2)
}
\arguments{
\item{series_matrix}{The matrix contains several rows to be caculated distance.}

\item{r1}{The weight parameter of the absolute distance between two series.}

\item{r2}{The weight parameter of the stability of the difference of two series. r1 add r2 must be 1.}
}
\value{
The distance matrix of every pair of rows in series_matrix. The class is \code{\link[stats]{dist}}.
}
\description{
This function caculates enhanced Euclidean distance for a given matrix.
}
\examples{
enhanced_euclidean(matrix(runif(100,0,100),20,50))

}
