% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_evaluate.R
\name{flow_evaluate}
\alias{flow_evaluate}
\title{Evaluate the forecast result.}
\usage{
flow_evaluate(rv, pv, method = "all")
}
\arguments{
\item{rv}{real value of traffic flow, a row from a dataframe;}

\item{pv}{forecast value, can be the return value of flow_knn function,
a row from a dataframe;}

\item{method}{can be "mse", "rmse", "mae" or "all"}
}
\description{
This function evaluates the forecast result in several evaluation criterions.
}
