% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_forecastplot.R
\name{flow_forecastplot}
\alias{flow_forecastplot}
\title{Plot forecast result.}
\usage{
flow_forecastplot(rv, pv)
}
\arguments{
\item{rv}{real value of traffic flow, a row from a dataframe;}

\item{pv}{forecast value, can be the return value of flow_knn function,
a row from a dataframe;}
}
\description{
Make plot to show the result of flow forecast as ggplot2 style.
}
\details{
Visualize the result of forecast is helpful to check the efficiency
of forecast. This function use ggplot2 to make the plot more elegant.
}
\section{Two Components}{

\enumerate{
    \item the real value of traffic flow, marked as blue point;
    \item the forecast value of traffic flow, marked as red line;
}
}

\seealso{
\code{\link[ggplot2]{ggplot2-package}}
}
