% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_loess.R
\name{handle_loess}
\alias{handle_loess}
\title{Caculate LOESS value from a dataframe.}
\usage{
handle_loess(df, exp, resp, sp)
}
\arguments{
\item{df}{The dataFrame contains series to be loess.}

\item{exp}{The sequence column number of explaining variable,
which is a column of df.}

\item{resp}{The sequence column number of response variable,
which is another column of df.}

\item{sp}{Span parameter of loess() function.}
}
\description{
This function caculates LOESS(locally weighted regression) by given a dataframe.
}
\details{
LOESS (locally weighted regression) is a technique which can smooth a series. It
needs a explaining variable as exp, a response variable as resp, a parameter span.
By using loess() function in stats package, this function can caculate LOESS value
for a series from a dataframe.
}
\seealso{
\code{\link[stats]{loess}}
}
