% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_loess.R
\name{handle_loess_fordf}
\alias{handle_loess_fordf}
\title{Caculate LOESS value from a dataframe.}
\usage{
handle_loess_fordf(df, sp)
}
\arguments{
\item{df}{The dataFrame contains series to be loess.}

\item{sp}{Span parameter of loess() function.}
}
\description{
This function caculates LOESS (locally weighted regression) by given a dataframe.
}
\details{
LOESS (locally weighted regression) is a technique which can smooth a series. It
needs a explaining variable as exp, a response variable as resp, a parameter span.
By using loess() function in stats package, this function can caculate LOESS value
for every row in the input dataframe.
}
\seealso{
\code{\link[stats]{loess}}
}
